mqtt_publisher_app/pyproject.toml
[tool.poetry]
name = "mqtt-publisher-app"
version = "0.1.0"
description = ""
authors = ["Sheena <11192150+sheenacodes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
publish = "mqtt_publisher_app:__init__.main"
